# packages/core/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Computer Interaction API",
    "version": "1.0.0",
    "description": "API to interact with and control a computer.",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4242/{computer_name}",
      "description": "Local development server",
      "variables": {
        "computer_name": {
          "default": "your-computer-name",
          "description": "Computer name"
        }
      }
    }
  ],
...
  "paths": {
...
    "/bash/command": {
      "post": {
        "tags": ["Bash"],
        "summary": "Execute Bash Command",
        "description": "Executes a given bash command and returns the output. If the command runs longer than 15 seconds, it will be moved to background and return a process ID.",
        "operationId": "bashExecuteCommand",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed successfully or moved to background.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/bash/process/{pid}/output": {
      "get": {
        "tags": ["Bash"],
        "summary": "Get Process Output",
        "description": "Gets the output of a background process by its PID.",
        "operationId": "bashGetProcessOutput",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "description": "Process ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Process output retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "404": {
            "description": "Process not found"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/bash/process/{pid}": {
      "delete": {
        "tags": ["Bash"],
        "summary": "Terminate Process",
        "description": "Terminates a background process by its PID.",
        "operationId": "bashTerminateProcess",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "description": "Process ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Process terminated successfully"
          },
          "404": {
            "description": "Process not found"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/bash/processes/managed": {
      "get": {
        "tags": ["Bash"],
        "summary": "List Managed Processes",
        "description": "Lists all background processes managed by the API.",
        "operationId": "bashListManagedProcesses",
        "responses": {
          "200": {
            "description": "List of managed processes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessListResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/bash/restart": {
      "post": {
        "tags": ["Bash"],
        "summary": "Restart System",
        "description": "Restarts the system.",
        "operationId": "bashRestartSystem",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestartParams"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "System restart initiated"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPError": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
...
      "CommandParams": {
        "type": "object",
        "properties": {
          "command": {
            "type": "string",
            "description": "Bash command to execute."
          }
        },
        "required": ["command"]
      },
      "CommandResponse": {
        "type": "object",
        "properties": {
          "output": {
            "type": "string",
            "description": "Command output if completed.",
            "nullable": true
          },
          "process_id": {
            "type": "integer",
            "description": "Process ID if running in background.",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "Status of the command execution.",
            "enum": ["completed", "background"]
          }
        },
        "required": ["status"]
      },
      "ProcessInfo": {
        "type": "object",
        "properties": {
          "pid": {
            "type": "integer",
            "description": "Process ID"
          },
          "command": {
            "type": "string",
            "description": "Command that started the process"
          },
          "status": {
            "type": "string",
            "description": "Current status of the process"
          },
          "cpu_percent": {
            "type": "number",
            "description": "CPU usage percentage",
            "nullable": true
          },
          "memory_percent": {
            "type": "number",
            "description": "Memory usage percentage",
            "nullable": true
          },
          "create_time": {
            "type": "number",
            "description": "Process creation time (Unix timestamp)",
            "nullable": true
          }
        },
        "required": ["pid", "command", "status"]
      },
      "ProcessListResponse": {
        "type": "object",
        "properties": {
          "processes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProcessInfo"
            },
            "description": "List of processes"
          }
        },
        "required": ["processes"]
      },
      "RestartParams": {
        "type": "object",
        "description": "No parameters needed for restart"
      }
    },
    "responses": {
      "InternalServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/HTTPError"
            }
          }
        }
      },
...
    },
...
  }
}